// basta-back/src/common/enums/post.enum.ts
export enum PostStatus {
  NaoIniciado = 'nao_iniciado',
  Captando = 'captando',
  Editando = 'editando',
  Criando = 'criando',
  Aprovacao = 'aprovacao',
  CopyCapa = 'copy_capa',
  EmPublicacao = 'em_publicacao',
  Publicado = 'publicado',
}

export enum PostType {
  Criativo = 'criativo',
  Video = 'video',
}

export enum PostFormat {
  Feed = 'feed',
  Story = 'story',
  Reels = 'reels',
  Capa = 'capa',
}

export enum Platform {
  Instagram = 'instagram',
  WhatsApp = 'whatsapp',
  YouTube = 'youtube',
}